Implement Pydantic for `Job`
